Online Payment Fraud Detection Project New
Business Understanding - The objective of this project is to predict Online Payment Fraud using Lachine Learning models depending on factors presented. It is a binary classification task to predict if online payments are fraudulent or non-fraudulent.
Data Understanding - The data provided is in a csv file containing 1048575 records from Blossom Bank. 
Modelling - 3 models will be evaluated for this classification task: logistic Regression, Decision Tree and Random Forest. Training and test data was split 60:40.
Evaluation - Each of the 3 models will be compared and evaluated